fix: height sentinel, zero-column row walk, and a nullable-PK hazard
Three unrelated defects that share a root: a value that looks safe until
it crosses the u64 height boundary or the snapshot's ordering contract.

`orchestrator/main.rs` passed `u64::MAX` to the validator debug view as
"any height >= tip". Heights map onto SQLite INTEGER by lossless
bit-cast, so that arrives as -1, and the CTE filters
`effective_height <= ?` where every height is >= 0 — the probe matched
zero rows unconditionally. `wait_for_formation` therefore burned its full
120 s on EVERY mesh creation and reported 0 seated validators. Now
`i64::MAX as u64`, which `src/db/consensus.rs` already establishes as the
safe ceiling and tests at that boundary.

This is the same inversion the master-merge commit reasoned about
explicitly for hopnet-mount's `ANCHOR_INIT` and then missed one file
over. It does NOT invalidate the container scenario results already
reported: the call site is documented "Non-fatal: tests do their own
baseline waits", and the scenarios' own assertions (heights advancing,
replica hashes agreeing, exit codes) never consult it.

`read_table_header` now rejects a unit that declares rows but no columns.
Such a unit makes `walk_table_rows` advance the cursor by nothing per
row, so `read_bytes` never reaches the end and never errors: a patched
`row_count` spins up to 2^64 times consuming no input. `import_section`
compares columns before walking, but the PARSE-TO-SKIP branch
deliberately does not — skipping an unknown section IS a full structural
decode of bytes whose shape is not trusted, which is precisely where a
hostile artifact would aim. Not reachable in production (both importers
verify the blake3 hash first, so it needs a preimage); this is the
routine's own bounds check rather than a defence something else provides.

For the record, the review's framing of this one was wrong:
`walk_table_rows` is the importer's row decoder, not a paginated
serializer walk. It has no keyset cursor and no primary-key involvement,
the serializer's walk is a single streaming `SELECT ... ORDER BY` with no
LIMIT, a table with no PK is a hard `MissingPrimaryKey` error caught in
CI by `empty_db_golden`, and none of the three paths named even reach the
importer. The zero-column case is the one real hole in the area.

`photo_ingress_responsibility` gets a test pinning
`idx_ingress_resp_personal` as load-bearing for CONSENSUS. Its PK is
`(user_id, library_id)` with `library_id` nullable, and SQLite treats
NULLs as distinct inside a PRIMARY KEY — so without that partial index a
user could hold two personal-scope rows and the canonical
`ORDER BY "library_id","user_id"` would stop being a total order. Two
honest replicas could then serialize the same state in different row
orders and compute different section hashes. Since the table is exported
in the photos snapshot section (declared during the master merge), that
is a divergence at the seal, where every validator recomputes the
artifact hash to vote — not a query returning rows oddly. The index reads
as redundant with the PK; it is not, and now cannot be removed quietly.

Evidence: `import_rejects_a_zero_column_unit` (29 hopnet-common tests),
`personal_scope_index_keeps_the_exported_row_order_total` (27
hopnet-photos tests), workspace sweep, clippy -D warnings and fmt green.

+++ orchestrator/main.rs
@@ -569,7 +569,14 @@ async fn wait_for_formation(
         if let Ok(resp) = client
             .post(&url)
             .header("Authorization", format!("Bearer {}", token))
-            .json(&u64::MAX) // any height >= tip resolves to the latest set
+            // `i64::MAX`, not `u64::MAX`: heights map onto SQLite INTEGER
+            // by lossless bit-cast, so anything above i64::MAX lands
+            // NEGATIVE and the validator CTE's `effective_height <= ?`
+            // matches ZERO rows — every height is >= 0. With u64::MAX this
+            // probe never saw a validator, so it burned its full timeout on
+            // every mesh creation and reported 0 seated.
+            // `src/db/consensus.rs` pins i64::MAX as the safe ceiling.
+            .json(&(i64::MAX as u64)) // any height >= tip resolves to the latest set
             .timeout(std::time::Duration::from_secs(5))
             .send()
             .await
